+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * Add support for the new talent tree system [#105]
 * Add option to automatically bind unassigned modifier keys to cast a binding
   * Creating a binding for 'Q' can automatically also bind 'SHIFT-Q', 'ALT-Q', etc.
+* Add tooltip explaining search filters to search box.
 
 ### Changed